Implement HKD_ADD_EVENT_HANDLER ioctl
The hyperkernel driver registers a miscdevice through the linux kernel
that can be open'd with "/dev/hkd".

HKD implements the HKD_ADD_EVENT_HANDLER ioctl. The argument to this ioctl
is the address of a 'struct hkd_event_handler'. Userspace must set the
pid and eventfd fields value prior to ioctling so that the driver
can properly notify it.

On success, the event handler will have the physical vector
and the Linux IRQ number of the event

Signed-off-by: Connor Davis <davisc@ainfosec.com>