Pass pci=nobar to prevent the PCI gap problem
Linux expects BIOS to carve out a gap in the E820 to which it
may assign to the PCI subsystem, but if pci=nobar is passed on
the command line it seems to be OK with one without gaps.

Note that with this change we no longer pass-through writes to
the pass-through device's bridge.

Signed-off-by: Connor Davis <davisc@ainfosec.com>